feat: implement EVPN Type 5 path construction in GoBGP runtime
Replaces the ErrEVPNNotImplemented stub with a real implementation that
builds and advertises EVPN Type 5 IP Prefix routes for each SRv6
endpoint prefix in a BGPAdvertisement. The route distinguisher is
derived from the BGPRouter's routerID (Type 1 IP-address:0), the
MpReachNLRI next-hop is the node's primary IPv6 address, and route
target communities are parsed from the advertisement's communities
field. Withdrawal is also supported via b.DeletePath.

Also adds a configurable BGP_LISTEN_PORT environment variable to
galactic-router so the tenant GoBGP instance can bind on port 1790
(port 179 is occupied by the FRR underlay on each worker node).

+++ internal/cni/cni_test.go
@@ -25,7 +25,7 @@ func fakeClient(objs ...client.Object) client.Client {
 }
 
 // routerForNode builds a BGPRouter with spec.targetRef.name set to nodeName.
-func routerForNode(name, nodeName, namespace string, asn uint32) *bgpv1alpha1.BGPRouter {
+func routerForNode(name, nodeName, namespace string, asn int64) *bgpv1alpha1.BGPRouter {
 	return &bgpv1alpha1.BGPRouter{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      name,
